Extracted options for KeyManager, KeyManagerFactory, TrustManager, TrustManagerFactory
